Add files via upload
Design and implement a C++ class to represent a person, which contains a name and a birthday at minimum.  Additionally, design and implement a C++ class to represent an ADT Dictionary of Person objects.  Do not use an existing library class for the dictionary, you must implement your own.

Searching for a person in your dictionary must be an average-case O(log n) operation; adding and removing can be O(n); and the "list everyone who was born in a given month" operation can be an O(n) operation.